CHKN2-14 User클래스, schema cardNumber Integer로 모두 변환.

+++ src/main/java/com/checkin/CheckIn/service/CheckInOutService.java
@@ -17,7 +17,7 @@ public class CheckInOutService {
     private final UserMapper userMapper;
 
     @Transactional
-    public UserResponseDto checkInService(Long cardNumber) {
+    public UserResponseDto checkInService(Integer cardNumber) {
         return UserResponseDto.builder()
                 .cardNumber(cardNumber)
                 .userCursus("42Cursus")
@@ -28,7 +28,7 @@ public class CheckInOutService {
     }
 
     @Transactional
-    public void checkOutService(Long cardNumber) {
+    public void checkOutService(Integer cardNumber) {
 
     }
 }
