+++ src/main/resources/db/schema.sql
@@ -4,7 +4,7 @@ CREATE TABLE USER (
                       id              BIGINT          PRIMARY KEY AUTO_INCREMENT,
                       intra_id        BIGINT          NOT NULL,
                       username        VARCHAR(255)    NOT NULL,
-                      card_number     SMALLINT        UNIQUE,
+                      card_number     INTEGER        UNIQUE,
                       check_in        DATETIME        ,
                       check_out       DATETIME
 );
